refactor: move getLastMci to db.ts
Pairs it with updateLastMci; replay keeps only the last_mci global.
Frees the replay test from mocking providers.
